requirements.txt
PyGithub==1.77
requests==2.31.0
python-dateutil==2.8.2
pytz==2024.4
pyyaml==6.0
Pygments==2.17.2
